refactor CLI operations into separate module
